nginx/html/adsuminternational.net/CreateUser/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
fido2==1.1.2
requests==2.26.0
python-dotenv==1.0.0